Move Gather shape inference into a submodule
Move the Gather, GatherElements and GatherND operators, plus their tests,
from ops.rs into a new ops/gather.rs submodule.
